Error caught when a correct key is not being handled and API calls have been modified to handle errors.
