--- lib/spotify_interface_web/controllers/page_controller.ex
@@ -1,9 +1,0 @@
-defmodule SpotifyInterfaceWeb.PageController do
-  use SpotifyInterfaceWeb, :controller
-
-  def home(conn, _params) do
-    # The home page is often custom made,
-    # so skip the default app layout.
-    render(conn, :home, layout: false)
-  end
-end
