--- lib/spotify_interface_web/controllers/page_html.ex
@@ -1,5 +1,0 @@
-defmodule SpotifyInterfaceWeb.PageHTML do
-  use SpotifyInterfaceWeb, :html
-
-  embed_templates "page_html/*"
-end
