+++ lib/spotify_interface_web/live/home/home_live.html.heex
@@ -1,3 +1,4 @@
+<.flash kind={:info} flash={@flash} />
 <div class="flex w-full bg-[#121212]">
   <div class="w-full h-screen text-gray-200">
     <div class="flex flex-col">
